Unbreak Win64 build.
TBR=danno@chromium.org

Review URL: https://codereview.chromium.org/12095033

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@13535 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/x64/deoptimizer-x64.cc
@@ -500,8 +500,8 @@ void Deoptimizer::DoCompiledStubFrame(TranslationIterator* iterator,
   unsigned input_frame_size = input_->GetFrameSize();
 
   // JSFunction continuation
-  intptr_t input_frame_offset = input_frame_size - kPointerSize;
-  intptr_t output_frame_offset = output_frame_size - kPointerSize;
+  unsigned input_frame_offset = input_frame_size - kPointerSize;
+  unsigned output_frame_offset = output_frame_size - kPointerSize;
   intptr_t value = input_->GetFrameSlot(input_frame_offset);
   output_frame->SetFrameSlot(output_frame_offset, value);
